refactor DataStructureCVisitor
