+++ src/main/java/nl/ou/debm/common/feature2/DataStructureCVisitor.java
@@ -10,8 +10,6 @@ import org.antlr.v4.runtime.misc.Interval;
 import org.antlr.v4.runtime.tree.ParseTree;
 import org.antlr.v4.runtime.tree.TerminalNode;
 
-import javax.swing.plaf.synth.SynthTextAreaUI;
-import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -20,160 +18,16 @@ import java.util.Optional;
 // This visitor extracts information about testcases from C code. A testcase here means a DatastructureCodeMarker. The result of the visiting operation is the array recovered_testcases, which contains partially unprocessed information. For example, the codemarker is stored as string and not yet parsed. This is because the purpose of the DataStructureCVisitor is to extract information that can then be further processed.
 public class DataStructureCVisitor extends CBaseVisitor<Object>
 {
-    //
-    //  Classes used in this namespace
-    //
-
-    enum EScope {
-        global, local, functionParameter, forDeclaration
-    }
-
-    static class Testcase {
-        enum Status {
-            ok, variableNotFound
-        }
-        Status status;
-        DataStructureCodeMarker codemarker;
-        String variableAddressExpr; //variable address expression created by the decompiler. This is the second argument of a DataStructureCodeMarker function call.
-        VariableInfo varInfo;
-    }
-
-    // This is how you create a union type in java. A "NameInfo" object can hold either a VariableInfo or TypeInfo value.
-    static sealed class NameInfo permits VariableInfo, TypeInfo {
-        public String name;
-        public EScope scope;
-
-        public String toString() {
-            var sb = new StringBuilder();
-            sb.append("name: " + name + "\n");
-
-            String kind;
-            if (this instanceof VariableInfo) kind = "VariableInfo";
-            else                              kind = "TypeInfo";
-            sb.append("kind: " + kind + "\n");
-
-            sb.append("scope: " + scope + "\n");
-
-            TypeInfo typeInfo;
-            if (this instanceof VariableInfo casted) {
-                typeInfo = casted.typeInfo;
-            } else {
-                typeInfo = (TypeInfo)this;
-            }
-            sb.append("typeInfo.strType: " + typeInfo.strType + "\n");
-            sb.append("typeInfo.isPointer: " + typeInfo.isPointer);
-
-            return sb.toString();
-        }
-    }
-
-
-    // Data class for an intermediate step in the processing of decompiled code. At this point, the name and type of a variable are known, but the type specifier is not yet parsed. It is raw code that specifies a type, for example "unsigned int", "struct {int i;}", "struct name" or just "name" if name is a valid typedef. Anything that specifies a type in C.
-    static final class TypeInfo extends NameInfo {
-        public String strType;
-        public boolean isPointer;
-    }
-
-    static final class VariableInfo extends NameInfo {
-        public TypeInfo typeInfo = new TypeInfo();
-    }
-
-    // purpose: contain data about all variables in some scope
-    static class ScopeNameInfo {
-        private ArrayList<NameInfo> names = new ArrayList<>();
-        private HashMap<String, Integer> indices = new HashMap<>(); //maps a variable name to its index in variables
-
-        public boolean contains(String name) {
-            var idx = indices.get(name);
-            return idx != null;
-        }
-        public NameInfo get(String name) {
-            var idx = indices.get(name);
-            if (idx == null) return null;
-            else             return names.get(idx);
-        }
-        public void add(NameInfo variable) {
-            names.add(variable);
-            var idx = names.size() - 1;
-            if (indices.containsKey(idx)) throw new RuntimeException("variable name " + variable.name + " occurs multiple times within the same scope. This is not supported."); //todo can this be somewhow tolerated?
-            indices.put(variable.name, idx);
-        }
-
-        //dont write to this list
-        public ArrayList<NameInfo> getNames() {
-            return names;
-        }
-    }
-
-    // generalization of VariableInfo for nested scopes. A block in C code creates a new scope, which is represented in this class by adding a new ScopeVariableInfo to a stack of scopes. The meaning of a variable name is always the most recent one, so to get more information about a name, the stack of scopes must be iterated in reverse order.
-    // Note: this class is used to keep track of which variables are in scope at any one time while traversing a parse tree. In that use case, it is constantly modified to remain up to date. This makes it easy to determine, for each code marker that is found, which variable it refers to, as it must be a variable that is currently in scope.
-    static class NestedNameInfo {
-        private ArrayList<ScopeNameInfo> scopeStack = new ArrayList<>(1); //not of type Stack because I need to iterate over it, but elements are only appended and popped.
-
-        public NestedNameInfo() {
-            //create an initial scope
-            addScope();
-        }
-        public NestedNameInfo(ScopeNameInfo initial_scope) {
-            scopeStack.add(initial_scope);
-        }
-
-        public boolean contains(String name) {
-            for (int i=scopeStack.size()-1; i>=0; i--) {
-                var scope = scopeStack.get(i);
-                if (scope.contains(name)) return true;
-            }
-            return false;
-        }
-        public NameInfo get(String name) {
-            for (int i=scopeStack.size()-1; i>=0; i--) {
-                var scope = scopeStack.get(i);
-                if (scope.contains(name)) return scope.get(name);
-            }
-            return null;
-        }
-        // adds to the latest scope
-        public void add(NameInfo nameInfo) {
-            var last_scope = scopeStack.get(scopeStack.size()-1);
-            last_scope.add(nameInfo);
-        }
-
-        // returns the newly added scope
-        public ScopeNameInfo addScope() {
-            var ret = new ScopeNameInfo();
-            scopeStack.add(ret);
-            return ret;
-        }
-        // removes and returns the top scope
-        public ScopeNameInfo popScope() {
-            assert scopeStack.size() > 0;
-            var ret = scopeStack.get(scopeStack.size() - 1);
-            scopeStack.remove(scopeStack.size() - 1);
-            return ret;
-        }
-        // same as popScope except the top scope is not removed
-        public ScopeNameInfo currentScope() {
-            assert scopeStack.size() > 0;
-            var ret = scopeStack.get(scopeStack.size() - 1);
-            return ret;
-        }
-    }
-
-
-    //
-    //  Data members
-    //
-    ScopeNameInfo globals = new ScopeNameInfo();
+    // Keeps track of which names are in scope at any moment during the traversal of the parse tree.
+    NameInfo nameInfo = new NameInfo();
     ArrayList<Testcase> recovered_testcases = new ArrayList<>();
-    ArrayList<String> structDeclarations; //todo does this only contain structs?
 
 
 
     //
-    //  Functions
+    //  Helper functions
     //
 
-    //todo generally useful
     // This is an alternative to ctx.getText, which returns code without spaces, which is often problematic.
     public static String originalCode(ParserRuleContext ctx) {
         int a = ctx.start.getStartIndex();
@@ -182,7 +36,7 @@ public class DataStructureCVisitor extends CBaseVisitor<Object>
         return ctx.start.getInputStream().getText(interval);
     }
 
-
+    //todo there are 4. the fourth is struct member declarations, which are different because they can be a bitfield
     // There are 3 kinds of declarations used in the C grammar: regular declarations, forDeclarations and parameterDeclarations. To avoid having to write 3 different handlers for those declarations, I convert them to a common format, which is the regular declaration. Conversion is easy, as a function parameter declaration is a special case of a normal declaration (it always declares one variable and does not initialize), except that it doesn't end with a semicolon, so reparsing with an added semicolon works. For forDeclarations the same can be done.
     static CParser.DeclarationContext toRegularDeclaration(ParserRuleContext ctx)
     {
@@ -213,7 +67,7 @@ public class DataStructureCVisitor extends CBaseVisitor<Object>
         return code.trim().replaceAll("\\s+", " ");
     }
 
-    static void parseStruct(CParser.StructOrUnionSpecifierContext ctx, NestedNameInfo nameInfo)
+    static void parseStruct(CParser.StructOrUnionSpecifierContext ctx, NameInfo nameInfo)
     {
         nameInfo.addScope();
         /*
@@ -232,7 +86,7 @@ public class DataStructureCVisitor extends CBaseVisitor<Object>
     //    :   declarationSpecifiers initDeclaratorList? ';'
     //    |   staticAssertDeclaration
     //    ;
-    static void parseDeclaration(ParserRuleContext ctx, NestedNameInfo dest, EScope scope)
+    static void parseDeclaration(ParserRuleContext ctx, NameInfo dest, NameInfo.EScope scope)
     {
         var declarationContext = toRegularDeclaration(ctx);
         var declarationSpecifiers = declarationContext.declarationSpecifiers();
@@ -266,13 +120,13 @@ public class DataStructureCVisitor extends CBaseVisitor<Object>
         {
             final int lastSpaceIdx = normalizedCode.lastIndexOf(' ');
 
-            var nameInfo = new TypeInfo();
-            nameInfo.name = normalizedCode.substring(lastSpaceIdx + 1).replace(";", "");
-            nameInfo.scope = scope;
-            nameInfo.strType = typeSpecifiers.get(0);
-            nameInfo.isPointer = normalizedCode.charAt(lastSpaceIdx - 1) == '*';
-            dest.add(nameInfo);
-            System.out.println("parsed as typedef: " + nameInfo);
+            var elt = new NameInfo.TypeInfo();
+            elt.name = normalizedCode.substring(lastSpaceIdx + 1).replace(";", "");
+            elt.scope = scope;
+            elt.strType = typeSpecifiers.get(0);
+            elt.isPointer = normalizedCode.charAt(lastSpaceIdx - 1) == '*';
+            dest.add(elt);
+            System.out.println("parsed as typedef: " + elt);
             return;
         }
 
@@ -314,7 +168,7 @@ public class DataStructureCVisitor extends CBaseVisitor<Object>
                     var isPointer = declarator.pointer() != null;
                     var variableName = declarator.directDeclarator().getText();
 
-                    var info = new VariableInfo();
+                    var info = new NameInfo.VariableInfo();
                     info.scope = scope;
                     info.name = variableName;
                     info.typeInfo.strType = baseTypeSpec;
@@ -326,7 +180,7 @@ public class DataStructureCVisitor extends CBaseVisitor<Object>
         }
         else if (variableNameBug != null) {
             // There was no declarator list because of the C grammar bug. Specific handling of that:
-            var info = new VariableInfo();
+            var info = new NameInfo.VariableInfo();
             info.scope = scope;
             info.name = variableNameBug;
             info.typeInfo.strType = baseTypeSpec;
@@ -335,110 +189,6 @@ public class DataStructureCVisitor extends CBaseVisitor<Object>
         }
 
         System.out.println("");
-    }
-
-
-    // finds globals
-    // This function is not called for declarations in functions, because visitFunctionDefinition doesn't call visitChildren.
-    public Object visitDeclaration(CParser.DeclarationContext ctx) {
-        System.out.println("global declaration found: " + originalCode(ctx));
-        parseDeclaration(ctx, globals, EScope.global);
-        return null;
-    }
-
-    @Override
-    public Object visitFunctionDefinition(CParser.FunctionDefinitionContext ctx) {
-        /*
-        Ghidra creates empty structs, with even incorrect C code.
-        ANTLR sees these lines as function definitions.
-        Therefore, we return when no function body is found
-        */
-        if (ctx.compoundStatement() == null || ctx.compoundStatement().blockItemList() == null)
-            return null;
-
-        var nameInfo = new NestedNameInfo(globals); //start with the global scope because function scope includes the global scope
-        nameInfo.addScope(); //scope for this function
-
-        // parse the function parameters
-        // I only visit the declarator to be safe, because maybe the function contains a function declaration somewhere that has its own parameters. I'm not sure if that would be a problem or not.
-        (new CBaseVisitor<Void>() {
-            @Override
-            public Void visitParameterTypeList(CParser.ParameterTypeListContext ctx) {
-                var parameterDeclarations = Optional.of(ctx)
-                    .map(x -> x.parameterList())
-                    .map(x -> x.parameterDeclaration())
-                    .orElse(null);
-                if (parameterDeclarations == null)
-                    return null;
-
-                for (var parameterDeclaration : parameterDeclarations) {
-                    parseDeclaration(parameterDeclaration, nameInfo, EScope.functionParameter);
-                }
-                return null;
-            }
-        }).visit(ctx.declarator());
-
-        // Traverse the function parse tree with a (sub)visitor
-        // This visitor finds all codemarkers, and keeps track of all names (type names and variable names) that are in scope, which is necessary to interpret the codemarkers.
-        var subvisitor = new CBaseVisitor<Void>() {
-            @Override
-            public Void visitStatement(CParser.StatementContext ctx)
-            {
-                // This function doesn't do much with the statements themselves. It just handles a few cases where scopes must be created before further traversal of the parse tree.
-
-                // A compound (block) statement needs its own scope.
-                var compoundStatement = ctx.compoundStatement();
-                if (compoundStatement != null) {
-                    nameInfo.addScope();
-                    visitChildren(ctx);
-                    nameInfo.popScope();
-                    return null;
-                }
-
-                // Check if this is a for loop that has a declaration. If so, the declaration gets its own scope. This is necessary if the for loop has a block statement that reuses a name, for example:
-                // for (int i=0; i<length; i++) {
-                //      int i = 10; //this doesn't affect the i defined above
-                // }
-                var forDeclaration = Optional.of(ctx)
-                    .map(x -> x.iterationStatement())
-                    .map(x -> x.forCondition())
-                    .map(x -> x.forDeclaration())
-                    .orElse(null);
-                if (forDeclaration != null) {
-                    nameInfo.addScope();
-                    parseDeclaration(forDeclaration, nameInfo, EScope.forDeclaration);
-
-                    // Continue parsing the for loop statement
-                    // Note: this also traverses the forDeclaration again. Although it will probably never be useful, if the forDeclaration somehow contains a codemarker, it will be found and it may even refer to a variable from that same declaration, because they've been added to the nameInfo at this point. (I was surprised to learn that C allows that. You can do things like "int i=10, j=i;")
-                    visitChildren(ctx);
-                    nameInfo.popScope();
-                    return null;
-                }
-
-                // Do the default for all other statements. This includes for loops without a declaration.
-                visitChildren(ctx);
-                return null;
-            }
-
-            @Override
-            public Void visitDeclaration(CParser.DeclarationContext ctx) {
-                parseDeclaration(ctx, nameInfo, EScope.local);
-                visitChildren(ctx);
-                return null;
-            }
-
-            // Codemarkers are function calls, and function calls are expressions, so visiting the expressions will find all codemarkers.
-            // Every codemarker that is found should have a second parameter that contains the name of a variable. Which variable this name refers to is lookup up in nameInfo, which contains information about all variables that are currently in scope.
-            @Override
-            public Void visitExpression(CParser.ExpressionContext ctx) {
-                var foundTestcases = findTestcasesInCode(originalCode(ctx), nameInfo);
-                recovered_testcases.addAll(foundTestcases);
-                return null;
-            }
-        };
-        subvisitor.visit(ctx);
-
-        return null;
     }
 
     // This requires a workaround because the C grammar we use doesn't allow for visiting all identifiers, which would have made this easy.
@@ -466,7 +216,7 @@ public class DataStructureCVisitor extends CBaseVisitor<Object>
     // After decompilation, some differences may occur. For example, the decompiled function call may have extra parameters. We want to have some tolerance for such mistakes, so we make as few assumptions about the structure of the C code as possible. What is assumed is the following:
     // - Code marker strings may occur anywhere. They are easily recognized by their characteristic prefix, which is highly unlikely to occur in code naturally.
     // - After a code marker string there is a comma (the separator for function call parameters) followed by an expression. This expression is the memory address of the variable being tested.
-    public static List<Testcase> findTestcasesInCode(String code, NestedNameInfo nameInfo) {
+    public static List<Testcase> findTestcasesInCode(String code, NameInfo nameInfo) {
         var ret = new ArrayList<Testcase>();
 
         //todo replace breaks with error handling or continue??
@@ -537,7 +287,7 @@ public class DataStructureCVisitor extends CBaseVisitor<Object>
 
             if (testcase.status == Testcase.Status.ok)
             {
-                testcase.varInfo = (VariableInfo) nameInfo.get(variableName); //todo cast is safe but should not be necessary?
+                testcase.varInfo = (NameInfo.VariableInfo) nameInfo.get(variableName);
             }
 
             ret.add(testcase);
@@ -545,5 +295,109 @@ public class DataStructureCVisitor extends CBaseVisitor<Object>
 
         return ret;
     }
-    //todo: gaat dit goed om met onverwachte situaties?
+
+
+    //
+    //  CBaseVisitor overrides
+    //
+
+    @Override
+    public Object visitDeclaration(CParser.DeclarationContext ctx)
+    {
+        System.out.println("global declaration found: " + originalCode(ctx));
+        //todo the scope can never be anything else because there is no context available here to determine the exact kind of scope. is that a problem? I do call parseDelcaration explicitly for function parameters and for declarations so I think it works well this way.
+        NameInfo.EScope scope;
+        if (nameInfo.stackSize() == 1) scope = NameInfo.EScope.global;
+        else                           scope = NameInfo.EScope.local;
+        parseDeclaration(ctx, nameInfo, scope);
+        return null;
+    }
+
+    @Override
+    public Object visitFunctionDefinition(CParser.FunctionDefinitionContext ctx)
+    {
+        /*
+        Ghidra creates empty structs, with even incorrect C code.
+        ANTLR sees these lines as function definitions.
+        Therefore, we return when no function body is found
+        */
+        if (ctx.compoundStatement() == null || ctx.compoundStatement().blockItemList() == null)
+            return null;
+
+        nameInfo.addScope(); //scope for this function
+
+        // parse the function parameters
+        // I only visit the declarator to be safe, because maybe the function contains a function declaration somewhere that has its own parameters. I'm not sure if that would be a problem or not.
+        (new CBaseVisitor<Void>() {
+            @Override
+            public Void visitParameterTypeList(CParser.ParameterTypeListContext ctx) {
+                var parameterDeclarations = Optional.of(ctx)
+                    .map(x -> x.parameterList())
+                    .map(x -> x.parameterDeclaration())
+                    .orElse(null);
+                if (parameterDeclarations == null)
+                    return null;
+
+                for (var parameterDeclaration : parameterDeclarations) {
+                    parseDeclaration(parameterDeclaration, nameInfo, NameInfo.EScope.functionParameter);
+                }
+                return null;
+            }
+        }).visit(ctx.declarator());
+
+        // Traverse the function parse tree with a (sub)visitor. //todo dit hoeft geen subvisitor te zijn
+        // This visitor finds all codemarkers, and keeps track of all names (type names and variable names) that are in scope, which is necessary to interpret the codemarkers.
+        visitChildren(ctx);
+
+        nameInfo.popScope();
+        return null;
+    }
+
+    @Override
+    public Void visitStatement(CParser.StatementContext ctx)
+    {
+        // This function doesn't do much with the statements themselves. It just handles a few cases where scopes must be created before further traversal of the parse tree.
+
+        // A compound (block) statement needs its own scope.
+        var compoundStatement = ctx.compoundStatement();
+        if (compoundStatement != null) {
+            nameInfo.addScope();
+            visitChildren(ctx);
+            nameInfo.popScope();
+            return null;
+        }
+
+        // Check if this is a for loop that has a declaration. If so, the declaration gets its own scope. This is necessary if the for loop has a block statement that reuses a name, for example:
+        // for (int i=0; i<length; i++) {
+        //      int i = 10; //this doesn't affect the i defined above
+        // }
+        var forDeclaration = Optional.of(ctx)
+            .map(x -> x.iterationStatement())
+            .map(x -> x.forCondition())
+            .map(x -> x.forDeclaration())
+            .orElse(null);
+        if (forDeclaration != null) {
+            nameInfo.addScope();
+            parseDeclaration(forDeclaration, nameInfo, NameInfo.EScope.forDeclaration);
+
+            // Continue parsing the for loop statement
+            // Note: this also traverses the forDeclaration again. Although it will probably never be useful, if the forDeclaration somehow contains a codemarker, it will be found and it may even refer to a variable from that same declaration, because they've been added to the nameInfo at this point. (I was surprised to learn that C allows that. You can do things like "int i=10, j=i;")
+            visitChildren(ctx);
+            nameInfo.popScope();
+            return null;
+        }
+
+        // Do the default for all other statements. This includes for loops without a declaration.
+        visitChildren(ctx);
+        return null;
+    }
+
+    // Codemarkers are function calls, and function calls are expressions, so visiting the expressions will find all codemarkers.
+    // Every codemarker that is found should have a second parameter that contains the name of a variable. Which variable this name refers to is lookup up in nameInfo, which contains information about all variables that are currently in scope.
+    @Override
+    public Void visitExpression(CParser.ExpressionContext ctx) {
+        var foundTestcases = findTestcasesInCode(originalCode(ctx), nameInfo);
+        recovered_testcases.addAll(foundTestcases);
+        return null;
+    }
 }
